Rearrange Arguments For REac[p] Commands
A little unnatural order of arguments existed for commands REac and
REacp to create analysis.

<prog_name> <cmdline_args> <ai_model>.

This resulted in commands like the following :
- REac xorprg "" binnet-0.5-x86-windows

That empty string in between looks quite ugly, and is only optional, but
due to the order of arguments it had to be passed like that. Now with
the reordering that "cmdline_args" parameter exists at the end of
command argument order and hence is not necessary to be passed. So the
same command can be executed as the following :
- REac binnet-0.5-x86-windows xorprg