Fix Filters Argument & Command Help Generation (#3)
* Fix Filters Argument & Command Help Generation

* Workflow Exec Everytime, Branch Selection, Plugin Version

The workflow now executes on every push and in any branch.
Earlier the workflow was building from master branch only, so modifications
in local branch wasn't getting built and tested. This is fixed now by
adding a `branchName` command line argument in build scripts.

We also have a better plugin version system now. This system is dynamic
and is generated automatically by the build system.

* Run Only On Push

* Fallback To Working Build Scripts

* Fix BranchName Issue

* Fix PluginVersion Header Path In CMakeLists.txt